[rename] : 카테고리 폴더 변경
